error

+++ routers/productRouter.js
@@ -1,87 +1,81 @@
 const { Router } = require("express");
-const { Product, Option } = require("../models");
-const orderService = require("../services/productService");
+const { Product, Option, User } = require("../models");
+const { authenticateUser, isAdmin } = require("../middleware/isAdmin");
+const productService = require("../services/productService");
 
 const productRouter = Router();
 
-productRouter.get('/', async (req, res, next) => {
-    //console.log("해당상품을 조회하였습니다.");
-
+productRouter.get("/", async (req, res, next) => {
     const { products, category } = req.query;
-    //console.log(`id : ${products}` );
 
     //const BestPList = products.sort({sales : -1}).limit(10);
-    //const NewPList = products.sort({entryDate : -1}).limit(10);
-    //const ReviewList = products.sort({entryDate : -1}).limit(10);
+    //const NewPList = products.sort({stockedAt : -1}).limit(10);
+    //const ReviewList = products.sort({stockedAt : -1}).limit(10);
 
-    if ((!category || typeof category !== 'string') && !products) {
-        const allProducts = await productService.getAllProducts();
-        return res.status(200).json({
-            //Bestproduct,
-            //Newproduct,
-            products: allProducts,
+    if (products !== undefined && products !== null) {
+        products.split(",").forEach((eachProduct) => {
+            if (!eachProduct instanceof String) {
+                const error = new Error("찾으려는 물품 값이 유효하지 않습니다.");
+                error.status = 400;
+                return next(error);
+            }
         });
     }
 
-    //카테고리가 존재하는 경우 카테고리를 기준으로 물품 데이터를 받아옴
-    if (category) {
-        try {
-            const filteredProducts = await productService.getProductsByCategory(category);
-            return res.json(filteredProducts);
+    if (category !== undefined && category !== null) {
+        if (!category instanceof String) {
+            const error = new Error("찾으려는 카테고리 값이 유효하지 않습니다.");
+            error.status = 400;
+            return next(error);
+        }
+    }
 
-        } catch (err) {
-            const error = new Error(" ");
+    //둘 다 입력
+    if (category !== undefined && category !== null && products !== undefined && products !== null) {
+        const arrOfProductId = products.split(",");
+        const productsInId = await productService.getProductsById(arrOfProductId);
+
+        const filteredProductByCategory = productsInId.filter((eachProduct) => {
+            return eachProduct.category.some((eachCategory) => {
+                return eachCategory.categoryName === category;
+            });
+        });
+
+        if (filteredProductByCategory.length === 0) {
+            const error = new Error("찾으려는 물품이 존재하지 않습니다.");
             error.status = 404;
             return next(error);
-        }   
-    }
-
-    if (products) {
-        const arrOfProductId = products.split(",");
-        const arrOfProductData = []; //Products 컬렉션 내 존재하는 물품 데이터
-        const arrOfNotExistProductId = []; //Products 컬렉션 내 존재하지 않는 요청 물품 데이터의 아이디
-        let eachProduct = 0;
-
-        for (eachProduct = 0; eachProduct < arrOfProductId.length; eachProduct++) {
-            try {
-                arrOfProductData.push(await productService.getProductById(arrOfProductId.at(eachProduct)));
-            } catch (error) {
-                arrOfNotExistProductId.push(arrOfProductId.at(eachProduct));
-            }
-        }
-
-        if (returnProducts.length > 0) {
-            //카테고리로 이미 반환받은 데이터가 있다면
-            //필터로 카테고리 내에 존재하는 물품을 반환
-            returnProducts = arrOfProductData.filter((product) => {
-                return product.category === category;
-            });
-        } else {
-            returnProducts = arrOfProductData;
-        }
-
-        if (arrOfNotExistProductId.length > 0) {
-            //존재하지 않는 물품의 아이디가 하나 이상일 때
-
-            //404는 에러이기 때문에 결과물을 담는건 x
-            //있는 product들만 담아서 응답으로 보내주도록 수정
-            return res.status(404).json({
-                products: returnProducts,
-                notExistProduct: arrOfNotExistProductId,
-            });
         } else {
             return res.status(200).json({
-                products: returnProducts,
+                products: filteredProductByCategory,
             });
         }
-    } else {
-        const allProducts = await productService.getAllProducts();
-        return res.json(200).json({
-            //Bestproduct,
-            //Newproduct,
-            products: allProducts,
+    }
+
+    //카테고리만 입력
+    if (category !== undefined && category !== null) {
+        const productsInCategory = await productService.getProductsByCategory(category);
+        return res.status(200).json({
+            products: productsInCategory,
         });
     }
+
+    //물품만 입력
+    if (products !== undefined && products !== null) {
+        const arrOfProductId = products.split(",");
+        const productsInId = await productService.getProductsById(arrOfProductId);
+        return res.status(200).json({
+            products: productsInId,
+        });
+    }
+
+    //카테고리, 물품 아이디 미 입력 시
+    const allProducts = await productService.getAllProducts();
+    return res.status(200).json({
+        //Best product,
+        //New product,
+        products: allProducts,
+    });
 });
 
 
@@ -101,14 +95,14 @@ productRouter.get('/', async (req, res, next) => {
 productRouter.post("/", async (req, res, next) => {
     //console.log("상품을 post합니다!");
     const { name, price, entryDate, discountRate, category, description, option, file } = req.body;
-    
+
     try {
         const existingProduct = await productService.checkProductExists(name);
-        
+
         if (existingProduct) {
             const error = new Error("이미 존재하는 상품입니다.");
-error.status = 409;
-throw error;
+            error.status = 409;
+            throw error;
 
         }
 
@@ -133,16 +127,58 @@ throw error;
     }
 });
 
-//상품 아이템 삭제
-productRouter.delete('/:id', async(req, res, next) => {
+//상품 수정 -> admin만 가능하게끔
+productRouter.patch('/:id', async (req, res, next) => {
     try {
-        console.log("삭제하는 라우터입니다.");
-        const id = req.params.id;
-        await Product.deleteOne({ _id: id });
-        //console.log("삭제 완료");
-        res.status(204).json({ message : '제품이 성공적으로 삭제되었습니다.' })
+        console.log("수정하는 라우터입니다.");
+        const productId = req.params;
+
+        const { name, price, sales, discountRate, category, description, option } = req.body;
+
+        const updatedProduct = await Product.findByIdAndUpdate(
+            productId,
+            {
+                name,
+                price,
+                sales,
+                discountRate,
+                category,
+                description,
+                option,
+            },
+            { new: true } //몽구스에서 지원하는 옵션 -> 업데이트 된 문서를 반환
+        );
+
+        if (!updatedProduct) {
+            return res.status(404).json({ message: '상품을 찾을 수 없습니다.' });
+        }
     } catch (err) {
         next(err);
+        return;
+    }
+})
+
+//상품 삭제 -> admin만 가능하게끔
+productRouter.delete('/:id', async (req, res, next) => {
+    try {
+        const id = req.params;
+        //const id = req.body.id;
+        if (id === undefined) {
+            res.status(404).json({ message: '해당 상품의 아이디가 필요합니다.' });
+        }
+        else {
+            const deleteProduct = await Product.deleteOne({ _id: id });
+
+            if (deleteProduct.deletedCount > 0) {
+                res.status(204).json({ message: '제품이 성공적으로 삭제되었습니다.' });
+            } else {
+                res.status(404).json({ message: '해당 ID의 상품을 찾을 수 없습니다.' });
+            }
+        }
+
+    } catch (err) {
+        next(err);
+        return;
     }
 })
